raspberry-edge/requirements.txt
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.0.4
backcall==0.2.0
beautifulsoup4==4.9.3
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorzero==1.1
contourpy==1.3.0
cycler==0.10.0
decorator==4.4.2
distro==1.5.0
Django==4.2.19
docutils==0.16
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
gpiozero==1.6.2
html5lib==1.1
idna==2.10
image==1.5.33
importlib_resources==6.5.2
ipython==7.20.0
ipython_genutils==0.2.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==3.1.5
jupyter-core==4.7.1
kiwisolver==1.4.7
logilab-common==1.8.1
MarkupSafe==3.0.2
matplotlib==3.9.4
mccabe==0.6.1
mock==4.0.3
mpmath==1.3.0
mypy-extensions==0.4.3
networkx==3.2.1
nudatus==0.0.5
numpy==1.19.5
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile==0.46
opencv-python==4.11.0.86
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
parso==0.8.1
pbr==5.5.0
pexpect==4.8.0
pickleshare==0.7.5
pigpio==1.78
pillow==11.1.0
prompt-toolkit==3.0.14
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.1
pyinotify==0.9.6
PyJWT==1.7.1
pyparsing==2.4.7
pyserial==3.5b0
python-dateutil==2.8.1
pytz==2025.1
PyYAML==6.0.2
requests==2.25.1
responses==0.12.1
roman==2.0.0
scipy==1.13.1
seaborn==0.13.2
semver==2.10.2
Send2Trash==1.6.0b1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.5.3
ssh-import-id==5.10
sympy==1.13.1
tflite-runtime==2.5.0.post1
toml==0.10.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.0.5
triton==3.2.0
typing-extensions==3.7.4.3
tzdata==2025.1
ufw==0.36
ultralytics==8.3.78
ultralytics-thop==2.0.14
urllib3==1.26.5
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
zipp==3.21.0
